Finally, some actual progress (1/4)
Ogg revorb (or as I've changed to calling it, regranularization) is now
mostly complete. The 'mostly' is because it remains largely untested; it
works fine for what files I have to test with, but I haven't tried any
'incorrect' files that might actually give it trouble. Next up, the
hardest thing of this whole project (probably): wem-to-ogg conversion.
Yay!