Rewrite makefiles from ground up
Focused on readability and usability; platform-determination and
variables are in `platform.mk`, build configuration and flags in
`config.mk`, vendor variables and recipes in `Makefile`.
(also, most variables are in snake_case now, also hopefully improving
readability).

This is hopefully, and I think it is, a large improvement from before;
enough so to increment the version to `0.0.2a`; 0.0.2 *official* will
come when I do some more actual codebase cleanup, bufixes, QOL stuff,
and when I update the readme.